[uka_sstp] refactor move examples/ -> tests/example.rs
