+++ .github/workflows/ci.yaml
@@ -55,29 +55,6 @@ jobs:
       - name: Cargo test
         run: cargo test
 
-  examples:
-    needs: build
-    runs-on: ubuntu-latest
-    env:
-      RUSTFLAGS: "-D warnings"
-      RUST_BACKTRACE: full
-      CARGO_INCREMENTAL: 0
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions-rs/toolchain@v1
-        with:
-          toolchain: stable
-          profile: minimal
-          override: true
-      - uses: Swatinem/rust-cache@v2
-        with:
-          shared-key: ubuntu-latest-stable20230820
-      - run: |
-          for example in */examples/*.rs
-          do
-              cargo run --example "$(basename "${example%.rs}")" -- $args
-          done
-
   lint:
     needs: build
     strategy:
